fix : 회원 api와 관리자 api 분리 및 명세 작성, 회원 정보 응답 시 프로필 이미지도 포함하도록 수정.
