+++ src/main/java/com/workhub/userTable/dto/user/response/UserListResponse.java
@@ -3,7 +3,9 @@ package com.workhub.userTable.dto.user.response;
 import com.workhub.userTable.entity.Status;
 import com.workhub.userTable.entity.UserRole;
 import com.workhub.userTable.entity.UserTable;
+import lombok.Builder;
 
+@Builder
 public record UserListResponse(
         Long userId,
         String userName,
@@ -12,18 +14,20 @@ public record UserListResponse(
         String phone,
         UserRole role,
         Status status,
+        String profileImg,
         Long companyId
 ){
     public static UserListResponse from(UserTable userTable){
-        return new UserListResponse(
-                userTable.getUserId(),
-                userTable.getUserName(),
-                userTable.getLoginId(),
-                userTable.getEmail(),
-                userTable.getPhone(),
-                userTable.getRole(),
-                userTable.getStatus(),
-                userTable.getCompanyId()
-        );
+        return UserListResponse.builder()
+                .userId(userTable.getUserId())
+                .userName(userTable.getUserName())
+                .loginId(userTable.getLoginId())
+                .email(userTable.getEmail())
+                .phone(userTable.getPhone())
+                .role(userTable.getRole())
+                .status(userTable.getStatus())
+                .profileImg(userTable.getProfileImg())
+                .companyId(userTable.getCompanyId())
+                .build();
     }
 }
