+++ src/test/java/com/workhub/userTable/service/UserServiceTest.java
@@ -4,10 +4,10 @@ import com.workhub.global.error.ErrorCode;
 import com.workhub.global.error.exception.BusinessException;
 import com.workhub.userTable.dto.user.response.UserDetailResponse;
 import com.workhub.userTable.dto.user.response.UserListResponse;
-import com.workhub.userTable.dto.user.response.UserLoginRecord;
+import com.workhub.userTable.dto.user.request.UserLoginRecord;
 import com.workhub.userTable.dto.user.request.AdminPasswordResetRequest;
 import com.workhub.userTable.dto.user.request.UserPasswordChangeRequest;
-import com.workhub.userTable.dto.user.response.UserRegisterRecord;
+import com.workhub.userTable.dto.user.request.UserRegisterRecord;
 import com.workhub.userTable.dto.user.response.UserTableResponse;
 import com.workhub.userTable.entity.Status;
 import com.workhub.userTable.entity.UserRole;
